refactor: extract PQValueLabel to separate file to break circular dependency
