+++ src/shared/components/plots/BoxScatterPlot.tsx
@@ -45,7 +45,7 @@ import classnames from 'classnames';
 import WindowStore from '../window/WindowStore';
 import LegendDataComponent from './LegendDataComponent';
 import LegendLabelComponent from './LegendLabelComponent';
-import { PQValueLabel } from 'shared/components/plots/MultipleCategoryBarPlot';
+import { PQValueLabel } from './PQValueLabel';
 import type { SampleIdsForPatientIds } from './PlotsTab';
 
 export interface IBaseBoxScatterPlotPoint {
